Move /src/react-date-select to /lib
