+++ src/samples/material/react-hook-form.tsx
@@ -1,6 +1,6 @@
 import { useForm, Controller } from "react-hook-form";
-import DateSelect from "../../react-date-select/DateSelect";
-import DateDropdown from "../../react-date-select/presets/material/DateDropdown";
+import DateSelect from "../../../lib/DateSelect";
+import DateDropdown from "../../../lib/presets/material/DateDropdown";
 
 type FormData = {
   date: string;
