+++ src/samples/vanilla/react-hook-form-with-original-render.tsx
@@ -1,6 +1,6 @@
 import { useForm, Controller } from "react-hook-form";
-import DateSelect from "../../react-date-select/DateSelect";
-import DateDropdown from "../../react-date-select/presets/vanilla/DateDropdown";
+import DateSelect from "../../../lib/DateSelect";
+import DateDropdown from "../../../lib/presets/vanilla/DateDropdown";
 
 type FormData = {
   date: string;
